+++ Changelog.md
@@ -16,6 +16,7 @@
 - AviSynth portable mode no longer uses soft links but moves files if necessary (stax76, [#473](/../../issues/473))
 - For x264 and x265 different code paths are used depending if the version name contains
   the keywords aMod or Asuna, Vanilla builds are supported as well (stax76)
+- Two audio bitrate issues were fixed (stax76)
 
 
 2.1.7.7 Beta (2021-02-05)